refactor(design-tokens): Consolidate `color` and `theme` attributes
