+++ packages/design-tokens/src/color/blendPoints.ts
@@ -24,11 +24,6 @@
 
  */
 
-export * from './breakpoints'
-export * from './easings'
-export * from './radii'
-export * from './shadows'
-export * from './size'
-export * from './space'
-export * from './transitions'
-export * from './typography'
+export type UIColorLevels = 1 | 2 | 3 | 4 | 5 | 6
+export const textBlends = [45, 65, 78, 88, 95, 99]
+export const uiBlends = [4, 12, 23, 34, 85]
