+++ packages/design-tokens/src/color/index.ts
@@ -24,24 +24,6 @@
 
  */
 
-import { colors } from '../../tokens/color'
-import { tintOrShadeUiColor } from './tintOrShadeUiColor'
-import { uiBlends } from './blend'
-
-const { background, text } = colors
-
-describe('tintOrShadeUiColor', () => {
-  describe('light (stock theme)', () => {
-    test('ui1', () =>
-      expect(tintOrShadeUiColor(uiBlends[0], background)).toEqual('#f4f4f4'))
-    test('ui5', () =>
-      expect(tintOrShadeUiColor(uiBlends[4], background)).toEqual('#262626'))
-  })
-
-  describe('dark-mode', () => {
-    test('ui1', () =>
-      expect(tintOrShadeUiColor(uiBlends[0], text)).toEqual('#33393f'))
-    test('ui5', () =>
-      expect(tintOrShadeUiColor(uiBlends[4], text)).toEqual('#fff'))
-  })
-})
+export * from './colors'
+export * from './types'
+export * from './utils'
