+++ packages/design-tokens/src/color/utils/index.ts
@@ -24,12 +24,9 @@
 
  */
 
-// Scales the amount mix percentage of two colors.
-
-// When mixing two colors, if one color has a low luminosity, dark background
-// colors for example, we need to mix at a higher percentage so the resulting
-// blended color is visually distinguishable.
-
-export const scaleMixAmount = (amount: number, number: number) => {
-  return amount * number
-}
+export { generateColors } from './generateColors'
+export { generateIntentShade } from './generateIntentShade'
+export { generatePressed, generateInteractive } from './generateStatefulColors'
+export { intentUIBlend } from './intentUIBlend'
+export { itemSelectedColor } from './itemSelectedColor'
+export { pickSpecifiableColors } from './pickSpecifiableColors'
