+++ packages/design-tokens/src/color/utils/mixScaledColors.test.ts
@@ -24,11 +24,22 @@
 
  */
 
-export * from './breakpoints'
-export * from './easings'
-export * from './radii'
-export * from './shadows'
-export * from './size'
-export * from './space'
-export * from './transitions'
-export * from './typography'
+import { uiBlends } from '../blendPoints'
+import { colors } from '../colors'
+import { mixScaledColors } from './mixScaledColors'
+
+const { background, text } = colors
+
+describe('mixScaledColors', () => {
+  test('default', () => {
+    expect(mixScaledColors(uiBlends[3], text, background)).toEqual('#b5b7b9')
+  })
+
+  test('dark mode', () => {
+    expect(mixScaledColors(uiBlends[3], background, text)).toEqual('#a3a6a8')
+  })
+
+  test('low but not super low luminance', () => {
+    expect(mixScaledColors(uiBlends[0], '#fff', '#555')).toEqual('#5d5d5d')
+  })
+})
