+++ packages/design-tokens/src/color/utils/scaleMixAmount.ts
@@ -24,11 +24,13 @@
 
  */
 
-export * from './breakpoints'
-export * from './easings'
-export * from './radii'
-export * from './shadows'
-export * from './size'
-export * from './space'
-export * from './transitions'
-export * from './typography'
+/**
+ * Scales the amount mix percentage of two colors.
+ *
+ * When mixing two colors, if one color has a low luminosity, dark background
+ * colors for example, we need to mix at a higher percentage so the resulting
+ * blended color is visually distinguishable.
+ */
+export const scaleMixAmount = (amount: number, number: number) => {
+  return amount * number
+}
