+++ packages/design-tokens/src/system/index.ts
@@ -72,7 +72,6 @@ export { Easings } from './easings'
 export { RadiusSizes, Radii } from './radii'
 export { Shadows } from './shadows'
 export * from './transitions'
-export * from './color'
 export * from './size'
 export * from './space'
 export * from './typography'
