+++ packages/design-tokens/src/theme/index.ts
@@ -24,11 +24,6 @@
 
  */
 
-export * from './breakpoints'
-export * from './easings'
-export * from './radii'
-export * from './shadows'
-export * from './size'
-export * from './space'
-export * from './transitions'
-export * from './typography'
+export * from './theme'
+export * from './utils'
+export type { ThemeCustomizations } from './ThemeCustomizations'
