+++ packages/design-tokens/src/theme/utils/index.ts
@@ -24,11 +24,4 @@
 
  */
 
-export * from './breakpoints'
-export * from './easings'
-export * from './radii'
-export * from './shadows'
-export * from './size'
-export * from './space'
-export * from './transitions'
-export * from './typography'
+export { generateTheme } from './generateTheme'
