+++ packages/design-tokens/src/utils/typography/googleFontUrl.ts
@@ -25,7 +25,6 @@
  */
 
 import { Theme } from '../../theme'
-
 interface GoogleFontSpecification {
   family: string
   /**
